Add files via upload
This code includes data cleaning and preprocessing, as well as the implementation of LSTM and ARIMA models. The predicted prices of the two are combined to be inout of an AI trader. Finally, a main_2() function is providing the readers a further try of another approach to the task. 
The code is structured using functions, to get advantage of the peculiar modularity of Python. 
Note: a draft() function is left at the end of the code, but useless.