Add files via upload
euklid_coding: data cleaning and preprocessing, ARIMA model, initial tries of LSTM. 

LSTMmodel: a file in which the work on LSTM model is began, before dividing the code for each of the datasets in a different notebook, as you can notice. 

Please refer to the main Method1 in which all the tries of these files are filtered into the final useful code.